CLN: Reduced the number of files to worry about for viewing specimen
