--- public/partials/single/main/dimensions/edit.php
@@ -1,44 +1,0 @@
-<?php
-
-function fossil_edit_dimensions( $fossil=null ) {
-    ?>
-
-    <div id="edit-fossil-dimensions" class="edit-fossil-popup">
-        <div class="edit-fossil">
-            <div class="edit-fossil-heading">
-                <h4>Dimensions</h4>
-            </div>
-            <div class="edit-fossil-body">
-                <form class="form">
-                    <div class="form-group" id="form-group-dimension-length">
-                        <label class="control-label">Length</label>
-                        <div class="input-group">
-                            <input class="form-control" type="text" id="edit-fossil-dimension-length" />
-                            <span class="input-group-addon">cm</span>
-                        </div>
-                    </div>
-                    <div class="form-group" id="form-group-dimension-width">
-                        <label class="control-label">Width</label>
-                        <div class="input-group">
-                            <input class="form-control" type="text" id="edit-fossil-dimension-width" />
-                            <span class="input-group-addon">cm</span>
-                        </div>
-                    </div>
-                    <div class="form-group" id="form-group-dimension-height">
-                        <label class="control-label">Height</label>
-                        <div class="input-group">
-                            <input class="form-control" type="text" id="edit-fossil-dimension-height" />
-                            <span class="input-group-addon">cm</span>
-                        </div>
-                    </div>
-
-                    <?php edit_comment_box( 'dimension' ); ?>
-                </form>
-            </div>
-            <div class="edit-fossil-footer">
-            </div>
-        </div>
-    </div>
-
-    <?php
-}
